Add custom logging configuration and communication interface for robot control
