+++ services/robot_control/app/custom_logging/logging_config.py
@@ -1,0 +1,15 @@
+import logging
+import os
+
+def setup_logger(log_file="custom_logging/logs/app.log", level=logging.DEBUG):
+    # Ensure the directory for the log file exists
+    os.makedirs(os.path.dirname(log_file), exist_ok=True)
+
+    logging.basicConfig(
+        level=level,
+        format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+        handlers=[
+            logging.FileHandler(log_file), 
+            logging.StreamHandler()
+        ]
+    )
